Clean up testing
